refactored RateLimitableFile to avoid creating a file handle
the server was reaching it's alloted file handle limit during a heavy unit test.
When a openning a bucket for the first time, the server creates an object representing
each file found. Part of creating each object involves openning a file handle to the
uploaded content. Refactored to lazily open the file when the content is actually
requested.