FIX response body when S3 object is not found.
The previous implementation returned empty response body if s3 object has not been found.
This caused the problems in AmazonS3Client (at least the java implementation) which was not able
to parse the response body because it expected the xml error structure.

This fix returns the response body XmlAdapter.error_no_such_key in that scenario.