s/max_keys/max-keys/ when parsing url query parameters
According to http://docs.aws.amazon.com/AmazonS3/latest/API/RESTBucketGET.html , the parameter is max-keys, not max_keys.
You need the conversion to an int as well, since you're gonna use max_keys as in int in the rest of the code. But I'm no rubyist, so I don't know if that conversion is idiomatic.