Fix incorrect variable when setting `custom_metadata`
The variable `metadata_struct` isn’t defined in `#create_metadata`,
which (obviously) would explode, though it wasn’t immediately clear
from the thrown exception and stacktrace, as it was something along
the lines of:

	ERROR NoMethodError: undefined method `md5' for nil:NilClass

Which is due to FileStore returning `nil` from `#store_object` because
`metadata_struct` is `nil` which leads to yet another NoMethodError
being thrown, etc, etc. This should fix that though.