Change redirect code from 307 to 303
This matches the behavior of the real S3.

The difference between 303 and 307 is significant. The HTTP 1.1 standard
says that a 307 indicates the request should be repeated using the same
method and post data, whereas a 303 always indicates a GET request.

This causes a bug when success_action_redirect is specified: a POST to
Fake S3 will subsequently POST to success_action_redirect instead of
GET as expected.