Support buckets with dots in them
In the v2 API, DELETE_BUCKET may be performed by passing the bucket name
as a subdomain. It appears the presence of dots may invoke this
behavior, so handling of the delete bucket operation needed updating
here too.

Because the v2 tests now pass the bucket by subdomain, extra subdomains
will probably need to be resolved locally via /etc/hosts for the tests
to pass. To reduce the number of such domains required, the delete
bucket test was updated to use the same bucket as the rest of the tests.